Fix issue: private message cannot be deleted.

+++ src/main/java/snw/mcpingbot/Main.java
@@ -47,7 +47,9 @@ public class Main extends BasePlugin {
         new JKookCommand("mcping")
                 .executesUser(
                         (sender, arguments, message) -> {
-                            message.delete();
+                            if (message instanceof TextChannelMessage) {
+                                message.delete();
+                            }
                             if (arguments.length == 0) {
                                 reply(sender, message, "您没有提供参数！");
                             } else {
